Cache server channels and roles consistently
This fixes a case where using Server#create_role would eventually cause
the role to be cached twice, once after the REST request, and then again
once we receive the subsequent GUILD_ROLE_CREATE event from the gateway.

This also changes Channels to be cached from the Server#create_channel
REST call, as we immediately know the channel exists and we may not ever
receive a gateway event that would trigger caching.