Use Role#update_from in ROLE_CREATE
In some cases, we may receive ROLE_CREATE for roles we already have
cached. Previously, this "new" role was added to the cache anyways, so
there would be duplicate role objects. Additionally, the handler now
does nothing if we receive an event for a guild we no longer have
cached.