Log current user, expected number of servers, number of loaded servers
This commit adds one line of logging, and promotes another line of
logging to INFO:

- When receiving READY from Discord, the currently connected user is
displayed as User#0000 ("distinct" style) as well as the *expected*
number of servers for the current shard. This will help users identify
that they have entered the correct bot token, mitigating developer error
of copying the wrong token, as well as giving the developer an insight
as to how many servers that shard is handling.

- Once all servers are loaded, *or* the timeout is reached, we log on
INFO how many servers we managed to load in that time. Combined with the
previous line of logging, this will give the developer an insight into
guild loading performance.